requirements.txt
streamlit
google-generativeai
python-dotenv 
supabase
